Fix warnings (cmake + compiler)(#11)
* cleanup ifdef for nvcc

* make eigen includes SYSTEM and ignore implicit fallthrough in cereal

* make cereal a SYSTEM include

* fix cmake warnings

* fix some implicit type conversion warnings

* undef max if defined

* add boost include for windows

* missing include

* something in mvsc defines CONST -_-

* msvc fixes

* gotta revert this, breaks nvcc 6.5

+++ include/ct/StringView.hpp
@@ -204,7 +204,7 @@ namespace ct
 
     constexpr int stoiRange(const char* start, const char* end) { return stoiImplRange(start, end); }
 
-    constexpr int stoiRange(const char* start, int len) { return stoiImplRange(start, start + len); }
+    constexpr int stoiRange(const char* start, size_t len) { return stoiImplRange(start, start + len); }
 
     ///// Implementation
     namespace detail
